# documentation/inventories/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuickOrderPreview": {
        "properties": {
          "ce_strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ce Strike"
          },
          "expiry": {
            "minLength": 1,
            "title": "Expiry",
            "type": "string"
          },
          "pe_strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Strike"
          },
          "underlying": {
            "minLength": 1,
            "title": "Underlying",
            "type": "string"
          },
          "underlying_exchange": {
            "default": "NSE_INDEX",
            "title": "Underlying Exchange",
            "type": "string"
          }
        },
        "required": [
          "underlying",
          "expiry"
        ],
        "title": "QuickOrderPreview",
        "type": "object"
      },
...
      "backend__routers__futures_risk__ConfigUpdate": {
        "properties": {
          "key": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          },
          "value": {
            "maxLength": 20000,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "ConfigUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Options Trading Platform for Indian Brokers",
    "title": "OpenBull",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/futures-risk/trades": {
      "get": {
        "description": "Return Futures-Risk trades for API-key clients.\n\nNinjaTrader uses this to restore one Bell drawing per phase after chart or\nworkspace reload. By default it returns the selected instrument's current\ntrading-session phases, which matches the Futures-Risk day-wise phase logic.",
        "operationId": "api_futures_risk_trade_list_api_v1_futures_risk_trades_get",
        "parameters": [
          {
            "in": "query",
            "name": "underlying",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Underlying"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "current_session",
            "required": false,
            "schema": {
              "default": true,
              "title": "Current Session",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(live|sandbox)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Futures Risk Trade List",
        "tags": [
          "api-v1"
        ]
      }
    },
    "/api/v1/futures-risk/trades/{trade_id}": {
      "get": {
        "description": "Return one Futures-Risk trade for API-key clients.",
        "operationId": "api_futures_risk_trade_detail_api_v1_futures_risk_trades__trade_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(live|sandbox)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Futures Risk Trade Detail",
        "tags": [
          "api-v1"
        ]
      }
    },
    "/api/v1/futures-risk/trades/{trade_id}/exit": {
      "post": {
        "description": "Close an active Futures-Risk trade from API-key clients.\n\nNinjaTrader uses this endpoint instead of placing an opposite quick order.\nThe shared Futures-Risk service still owns sandbox/live dispatch, broker\nconfirmation, phase completion, P&L, and event logging.",
        "operationId": "api_futures_risk_trade_exit_api_v1_futures_risk_trades__trade_id__exit_post",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Futures Risk Trade Exit",
        "tags": [
          "api-v1"
        ]
      }
    },
...
    "/web/fr/config": {
      "get": {
        "operationId": "get_config_web_fr_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "futures-risk"
        ]
      },
      "post": {
        "operationId": "update_config_web_fr_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__routers__futures_risk__ConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "futures-risk"
        ]
      }
    },
    "/web/fr/expiries": {
      "get": {
        "operationId": "expiries_web_fr_expiries_get",
        "parameters": [
          {
            "in": "query",
            "name": "underlying",
            "required": true,
            "schema": {
              "title": "Underlying",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exchange",
            "required": false,
            "schema": {
              "default": "NSE_INDEX",
              "title": "Exchange",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Expiries",
        "tags": [
          "futures-risk"
        ]
      }
    },
    "/web/fr/phases": {
      "get": {
        "description": "Phase history grouped by (underlying, session) with per-phase lifecycle\nsummary (entry/exit, P&L, duration, achieved targets, exit kind).",
        "operationId": "list_phases_web_fr_phases_get",
        "parameters": [
          {
            "in": "query",
            "name": "underlying",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Underlying"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(live|sandbox)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Phases",
        "tags": [
          "futures-risk"
        ]
      }
    },
    "/web/fr/quick-order-preview": {
      "post": {
        "description": "Session-auth quote preview for the browser quick-order popup.\n\nThe popup still prefers WebSocket ticks, but this endpoint gives it a\nREST/cache fallback for instruments whose broker stream is slow or missing.",
        "operationId": "quick_order_preview_web_fr_quick_order_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickOrderPreview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quick Order Preview",
        "tags": [
          "futures-risk"
        ]
      }
    },
...
  }
}